refactor(club-application): Shorten page.tsx
