+++ app/club-application/defaultClubData.js
@@ -1,0 +1,121 @@
+const defaultClubAdvisors = [
+    {
+      name: "",
+      email: "",
+      phoneNumber: "",
+    },
+  ];
+  
+  const defaultClubOfficers = [
+    {
+      role: "Club President",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: true,
+      isUsed: true,
+    },
+    {
+      role: "ICC Representative",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: true,
+      isUsed: true,
+    },
+    {
+      role: "Club Vice President",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: false,
+      isUsed: false,
+    },
+    {
+      role: "Club Secretary",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: false,
+      isUsed: false,
+    },
+    {
+      role: "Club Treasurer",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: false,
+      isUsed: false,
+    },
+    {
+      role: "Club Social Media Manager",
+      name: "",
+      email: "",
+      wNumber: "",
+      phoneNumber: "",
+      major: "",
+      gradeLevel: "",
+      isRequired: false,
+      isUsed: false,
+    },
+  ];
+  
+  const defaultMembers = [
+    {
+      name: "",
+      email: "",
+      wNumber: "",
+    },
+    {
+      name: "",
+      email: "",
+      wNumber: "",
+    },
+    {
+      name: "",
+      email: "",
+      wNumber: "",
+    },
+    {
+      name: "",
+      email: "",
+      wNumber: "",
+    },
+    {
+      name: "",
+      email: "",
+      wNumber: "",
+    },
+  ];
+  
+  const defaultClubAgreement = 
+    [
+      {
+        role: "Club President",
+        signature: "",
+        date: "",
+      },
+      {
+        role: "Club Advisor",
+        signature: "",
+        date: "",
+      },
+    ];
+
+
+export { defaultClubAdvisors, defaultClubOfficers, defaultMembers, defaultClubAgreement };
